+++ src/main/java/open/commons/spring/elastic/configuration/elasticsearch/DefaultClientBuilderConfiguration.java
@@ -32,7 +32,7 @@ package open.commons.spring.elastic.configuration.elasticsearch;
  * @version 0.1.0
  * @author Park Jun-Hong (parkjunhong77@gmail.com)
  */
-public class DefaultClientConfiguration extends ElasticsearchClientConfiguration<EndpointBuilderConfiguration, SccBuilderConfiguration, TccBuilderConfiguration> {
+public class DefaultClientBuilderConfiguration extends ElasticsearchClientBuilderConfiguration<EndpointBuilderConfiguration, SccBuilderConfiguration, TccBuilderConfiguration> {
 
     /**
      * <br>
@@ -49,7 +49,7 @@ public class DefaultClientConfiguration extends ElasticsearchClientConfiguration
      * @version 0.1.0
      * @author Park Jun-Hong (parkjunhong77@gmail.com)
      */
-    public DefaultClientConfiguration() {
+    public DefaultClientBuilderConfiguration() {
     }
 
     /**
@@ -58,7 +58,7 @@ public class DefaultClientConfiguration extends ElasticsearchClientConfiguration
      * @version 0.1.0
      * @author Park Jun-Hong (parkjunhong77@gmail.com)
      *
-     * @see open.commons.spring.elastic.configuration.elasticsearch.ElasticsearchClientConfiguration#setEndpoint(open.commons.spring.elastic.configuration.elasticsearch.EndpointBuilderConfiguration)
+     * @see open.commons.spring.elastic.configuration.elasticsearch.ElasticsearchClientBuilderConfiguration#setEndpoint(open.commons.spring.elastic.configuration.elasticsearch.EndpointBuilderConfiguration)
      */
     @Override
     public void setEndpoint(EndpointBuilderConfiguration endpoint) {
@@ -71,7 +71,7 @@ public class DefaultClientConfiguration extends ElasticsearchClientConfiguration
      * @version 0.1.0
      * @author Park Jun-Hong (parkjunhong77@gmail.com)
      *
-     * @see open.commons.spring.elastic.configuration.elasticsearch.ElasticsearchClientConfiguration#setSecure(open.commons.spring.elastic.configuration.elasticsearch.SccBuilderConfiguration)
+     * @see open.commons.spring.elastic.configuration.elasticsearch.ElasticsearchClientBuilderConfiguration#setSecure(open.commons.spring.elastic.configuration.elasticsearch.SccBuilderConfiguration)
      */
     @Override
     public void setSecure(SccBuilderConfiguration secure) {
@@ -84,7 +84,7 @@ public class DefaultClientConfiguration extends ElasticsearchClientConfiguration
      * @version 0.1.0
      * @author Park Jun-Hong (parkjunhong77@gmail.com)
      *
-     * @see open.commons.spring.elastic.configuration.elasticsearch.ElasticsearchClientConfiguration#setTerminal(open.commons.spring.elastic.configuration.elasticsearch.TccBuilderConfiguration)
+     * @see open.commons.spring.elastic.configuration.elasticsearch.ElasticsearchClientBuilderConfiguration#setTerminal(open.commons.spring.elastic.configuration.elasticsearch.TccBuilderConfiguration)
      */
     @Override
     public void setTerminal(TccBuilderConfiguration terminal) {